renamed interfaces to types; comple to es6; extracted plugin class out of index
