+++ tsconfig.json
@@ -3,7 +3,7 @@
     "outDir": "./dist",
     "allowJs": true,
     "module": "commonjs",
-    "target": "es5",
+    "target": "es6",
     "watch": false,
     "lib": [
       "es2015"
